Update version and description
For the description - there's no reason we can't support other 2.x
versions; we're just going to need different release branches.

Which ties into the version - the versioning scheme for this module
is now going to evolve in parallel with lunr.js itself.  The hope
is that if you're using lunr.js 2.x.y, lunr-mutable 2.x.z will
work for you